requirements.txt
groq
python-dotenv
gradio
google-generativeai 
pillow 
apscheduler
